Move UHCPlugin to core package
