+++ src/main/java/com/github/nthportal/uhc/commands/ConfCommandExecutor.java
@@ -1,6 +1,6 @@
 package com.github.nthportal.uhc.commands;
 
-import com.github.nthportal.uhc.UHCPlugin;
+import com.github.nthportal.uhc.core.UHCPlugin;
 import com.github.nthportal.uhc.core.Config;
 import com.google.common.base.Joiner;
 import org.bukkit.command.Command;
